v0.17.1c 令WaterMeshDeformer创建所有网格组件

+++ Assets/Resources_addressable/Scripts/Tilemap/Noise.cs
@@ -10,19 +10,19 @@ public class Noise{
         _noise_generator = new(seed);
     }
 
-    public float _perlin(int x, int y, float scale){
+    public float _perlin(float x, float y, float scale){
         _noise_generator.SetNoiseType(FastNoise.NoiseType.Perlin);
         _noise_generator.SetFrequency(0.01f*scale);
         float noise_value = _noise_generator.GetNoise(x, y);
         return noise_value;
     }
 
-    public float _perlin(int x, float scale){
+    public float _perlin(float x, float scale){
         float noise_value = _perlin(x, 0, scale);
         return noise_value;
     }
 
-    static float _perlin(int x, int y, float scale, int seed){
+    static float _perlin(float x, float y, float scale, int seed){
         FastNoise noise_generator = new(seed);
         noise_generator.SetNoiseType(FastNoise.NoiseType.Perlin);
         noise_generator.SetFrequency(0.01f*scale);
@@ -30,7 +30,7 @@ public class Noise{
         return noise_value;
     }
 
-    static float _perlin(int x, float scale, int seed){
+    static float _perlin(float x, float scale, int seed){
         float noise_value = _perlin(x, 0, scale, seed);
         return noise_value;
     }
